fixed editor middleware

+++ app/Http/Controllers/Backend/UsersController.php
@@ -24,74 +24,76 @@ class UsersController extends Controller
      */
     public function index(Request $request)
     {
-        if ($request->ajax()) {
-            $data = User::query()->latest();
-            return Datatables::of($data)
-                        ->addIndexColumn()
-                        ->editColumn('name',function($each){
-                            return Str::limit($each->name,45,'...');
-                        })
-                        ->editColumn('email',function($each){
-                            return Str::limit($each->email,45,'...');
-                        })
-                        ->editColumn('phone',function($each){
-                            return Str::limit($each->phone,45,'...');
-                        })
-                        ->editColumn('profile_photo',function($each){
-                            return '<div class="avatar avatar-lg avatar-4by3">
-                                        <img src="'.$each->profile_img_path().'" alt="" class="avatar-img rounded-circle">    
-                                    </div>';
-                        })
-                        ->editColumn('user_type',function($each){
-                            return $each->user_type ? config('const.user_type')[$each->user_type] : '-';
-                        })
-                        ->editColumn('user_agent',function($each){
-                            if ($each->user_agent){
-                                $agent = new Agent();
-                                $agent->setUserAgent($each->user_agent);
-                                $device = $agent->device();
-                                $platform = $agent->platform();
-                                $browser = $agent->browser();
-                
-                                return '<span class="badge badge-primary">'.$device.'</span></br>'.
-                                       '<span class="badge badge-success">'.$platform.'</span></br>'.
-                                       '<span class="badge badge-info">'.$browser.'</span>'.
-                                       '<span class="badge badge-primary">'.$each->ip.'</span>';
-                            }
-                            return '-';
-                        })
-                        
-                        ->editColumn('login_at',function($each){
-                            return $each->login_at ? Carbon::parse($each->login_at)->diffForHumans() : '-';
-                            
-                        })
-                        ->editColumn('updated_at',function($each){
-                            return Carbon::parse($each->created_at)->format('d-m-y H:i:s');
-                        })
-                        ->addColumn('action', function($each){
+        if (auth()->user()->user_type == 1) {
+            if ($request->ajax()) {
+                $data = User::query()->latest();
+                return Datatables::of($data)
+                    ->addIndexColumn()
+                    ->editColumn('name', function ($each) {
+                        return Str::limit($each->name, 45, '...');
+                    })
+                    ->editColumn('email', function ($each) {
+                        return Str::limit($each->email, 45, '...');
+                    })
+                    ->editColumn('phone', function ($each) {
+                        return Str::limit($each->phone, 45, '...');
+                    })
+                    ->editColumn('profile_photo', function ($each) {
+                        return '<div class="avatar avatar-lg avatar-4by3">
+                                            <img src="' . $each->profile_img_path() . '" alt="" class="avatar-img rounded-circle">    
+                                        </div>';
+                    })
+                    ->editColumn('user_type', function ($each) {
+                        return $each->user_type ? config('const.user_type')[$each->user_type] : '-';
+                    })
+                    ->editColumn('user_agent', function ($each) {
+                        if ($each->user_agent) {
+                            $agent = new Agent();
+                            $agent->setUserAgent($each->user_agent);
+                            $device = $agent->device();
+                            $platform = $agent->platform();
+                            $browser = $agent->browser();
 
-                            $info = '<a href="'. route('admin.users.show',$each->id) .'" type="button" class="btn btn-info btn-rounded">
-                                            <i class="material-icons">fullscreen</i>
-                                    </a>';
-                            $edit = '<a href="'. route('admin.users.edit',$each->id) .'" type="button" class="btn btn-light btn-rounded">
-                                            <i class="material-icons">edit</i>
-                                    </a>';
-                            if (Auth::id() == $each->id) {
-                                $delete = '';
-                            }else{
-                                $delete = '<a href="#" type="button" class="btn btn-danger btn-rounded delete" data-id="' . $each->id . '">
-                                            <i class="material-icons">close</i>
+                            return '<span class="badge badge-primary">' . $device . '</span></br>' .
+                                '<span class="badge badge-success">' . $platform . '</span></br>' .
+                                '<span class="badge badge-info">' . $browser . '</span>' .
+                                '<span class="badge badge-primary">' . $each->ip . '</span>';
+                        }
+                        return '-';
+                    })
+
+                    ->editColumn('login_at', function ($each) {
+                        return $each->login_at ? Carbon::parse($each->login_at)->diffForHumans() : '-';
+                    })
+                    ->editColumn('updated_at', function ($each) {
+                        return Carbon::parse($each->created_at)->format('d-m-y H:i:s');
+                    })
+                    ->addColumn('action', function ($each) {
+
+                        $info = '<a href="' . route('admin.users.show', $each->id) . '" type="button" class="btn btn-info btn-rounded">
+                                                <i class="material-icons">fullscreen</i>
                                         </a>';
-                            }
-                            
-                            $collect = $info.$edit.$delete;
-                            $action = '<div class="button-list">'.$collect.'</div>';
-                            return $action ;
-                        })
-                        ->rawColumns(['action','profile_photo','user_agent'])
-                        ->make(true);
+                        $edit = '<a href="' . route('admin.users.edit', $each->id) . '" type="button" class="btn btn-light btn-rounded">
+                                                <i class="material-icons">edit</i>
+                                        </a>';
+                        if (Auth::id() == $each->id) {
+                            $delete = '';
+                        } else {
+                            $delete = '<a href="#" type="button" class="btn btn-danger btn-rounded delete" data-id="' . $each->id . '">
+                                                <i class="material-icons">close</i>
+                                            </a>';
+                        }
+
+                        $collect = $info . $edit . $delete;
+                        $action = '<div class="button-list">' . $collect . '</div>';
+                        return $action;
+                    })
+                    ->rawColumns(['action', 'profile_photo', 'user_agent'])
+                    ->make(true);
+            }
+            return view('backend.users.index');
         }
-        return view('backend.users.index');
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -101,7 +103,10 @@ class UsersController extends Controller
      */
     public function create()
     {
-        return view('backend.users.create');
+        if (auth()->user()->user_type == 1) {
+            return view('backend.users.create');
+        }
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -112,37 +117,40 @@ class UsersController extends Controller
      */
     public function store(CreateUser $request)
     {
-        $profile_photo = null;
-        if ($request->hasFile('profile_photo')) {
-            $files = $request->file('profile_photo');
-            foreach($files as $file){
-                $extension = $file->getClientOriginalExtension();
-                $profile_photo = 'profile_'.Str::random(5)."-".date('his')."-". Str::random(3).".".$extension;
-                Storage::disk('public')->put('users/profile/'.$profile_photo, file_get_contents($file));
+        if (auth()->user()->user_type == 1) {
+            $profile_photo = null;
+            if ($request->hasFile('profile_photo')) {
+                $files = $request->file('profile_photo');
+                foreach ($files as $file) {
+                    $extension = $file->getClientOriginalExtension();
+                    $profile_photo = 'profile_' . Str::random(5) . "-" . date('his') . "-" . Str::random(3) . "." . $extension;
+                    Storage::disk('public')->put('users/profile/' . $profile_photo, file_get_contents($file));
+                }
             }
-        }
-        $cover_photo = null;
-        if ($request->hasFile('cover_photo')) {
-            $files = $request->file('cover_photo');
-            foreach($files as $file){
-                $extension = $file->getClientOriginalExtension();
-                $cover_photo = 'cover_'.Str::random(5)."-".date('his')."-". Str::random(3).".".$extension;
-                Storage::disk('public')->put('users/cover/'.$cover_photo, file_get_contents($file));
+            $cover_photo = null;
+            if ($request->hasFile('cover_photo')) {
+                $files = $request->file('cover_photo');
+                foreach ($files as $file) {
+                    $extension = $file->getClientOriginalExtension();
+                    $cover_photo = 'cover_' . Str::random(5) . "-" . date('his') . "-" . Str::random(3) . "." . $extension;
+                    Storage::disk('public')->put('users/cover/' . $cover_photo, file_get_contents($file));
+                }
             }
-        }
-        $data = new User();
-        $data->name = $request->name;
-        $data->email = $request->email;
-        $data->phone = $request->phone;
-        $data->address = $request->address;
-        $data->description = $request->description;
-        $data->profile_photo = $profile_photo;
-        $data->cover_photo = $cover_photo;
-        $data->user_type = $request->user_type;
-        $data->password = Hash::make('starfm@changeme');
-        $data->save();
+            $data = new User();
+            $data->name = $request->name;
+            $data->email = $request->email;
+            $data->phone = $request->phone;
+            $data->address = $request->address;
+            $data->description = $request->description;
+            $data->profile_photo = $profile_photo;
+            $data->cover_photo = $cover_photo;
+            $data->user_type = $request->user_type;
+            $data->password = Hash::make('starfm@changeme');
+            $data->save();
 
-        return redirect()->route('admin.users.index')->with('create', 'Created Successfully');
+            return redirect()->route('admin.users.index')->with('create', 'Created Successfully');
+        }
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -153,7 +161,10 @@ class UsersController extends Controller
      */
     public function show(User $user)
     {
-        return view('backend.users.edit', compact('user'));
+        if (auth()->user()->user_type == 1) {
+            return view('backend.users.edit', compact('user'));
+        }
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -164,7 +175,10 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
-        return view('backend.users.edit', compact('user'));
+        if (auth()->user()->user_type == 1) {
+            return view('backend.users.edit', compact('user'));
+        }
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -176,38 +190,41 @@ class UsersController extends Controller
      */
     public function update(UpdateUser $request, $id)
     {
-        $data = User::findOrFail($id);
-        $profile_photo = $data->profile_photo;
-        if ($request->hasFile('profile_photo')) {
-            Storage::disk('public')->delete('/users/profile/'.$profile_photo);
-            $files = $request->file('profile_photo');
-            foreach($files as $file){
-                $extension = $file->getClientOriginalExtension();
-                $profile_photo = 'profile_'.Str::random(5)."-".date('his')."-". Str::random(3).".".$extension;
-                Storage::disk('public')->put('/users/profile/'.$profile_photo, file_get_contents($file));
+        if (auth()->user()->user_type == 1) {
+            $data = User::findOrFail($id);
+            $profile_photo = $data->profile_photo;
+            if ($request->hasFile('profile_photo')) {
+                Storage::disk('public')->delete('/users/profile/' . $profile_photo);
+                $files = $request->file('profile_photo');
+                foreach ($files as $file) {
+                    $extension = $file->getClientOriginalExtension();
+                    $profile_photo = 'profile_' . Str::random(5) . "-" . date('his') . "-" . Str::random(3) . "." . $extension;
+                    Storage::disk('public')->put('/users/profile/' . $profile_photo, file_get_contents($file));
+                }
             }
-        }
-        $cover_photo = $data->cover_photo;
-        if ($request->hasFile('cover_photo')) {
-            Storage::disk('public')->delete('/users/cover/'.$cover_photo);
-            $files = $request->file('cover_photo');
-            foreach($files as $file){
-                $extension = $file->getClientOriginalExtension();
-                $cover_photo = Str::random(5)."-".date('his')."-". Str::random(3).".".$extension;
-                Storage::disk('public')->put('/users/cover/'.$cover_photo, file_get_contents($file));
+            $cover_photo = $data->cover_photo;
+            if ($request->hasFile('cover_photo')) {
+                Storage::disk('public')->delete('/users/cover/' . $cover_photo);
+                $files = $request->file('cover_photo');
+                foreach ($files as $file) {
+                    $extension = $file->getClientOriginalExtension();
+                    $cover_photo = Str::random(5) . "-" . date('his') . "-" . Str::random(3) . "." . $extension;
+                    Storage::disk('public')->put('/users/cover/' . $cover_photo, file_get_contents($file));
+                }
             }
-        }
-        $data->name = $request->name;
-        $data->email = $request->email;
-        $data->phone = $request->phone;
-        $data->address = $request->address;
-        $data->description = $request->description;
-        $data->profile_photo = $profile_photo;
-        $data->cover_photo = $cover_photo;
-        $data->user_type = $request->user_type;
-        $data->update();
+            $data->name = $request->name;
+            $data->email = $request->email;
+            $data->phone = $request->phone;
+            $data->address = $request->address;
+            $data->description = $request->description;
+            $data->profile_photo = $profile_photo;
+            $data->cover_photo = $cover_photo;
+            $data->user_type = $request->user_type;
+            $data->update();
 
-        return redirect()->route('admin.users.index')->with('update', 'Successfully Updated');
+            return redirect()->route('admin.users.index')->with('update', 'Successfully Updated');
+        }
+        abort(403, 'Unauthorized action.');
     }
 
     /**
@@ -218,9 +235,12 @@ class UsersController extends Controller
      */
     public function reset($id)
     {
-        $data = User::findOrFail($id);
-        $data->password = Hash::make('starfm@changeme');
-        $data->update();
+        if (auth()->user()->user_type == 1) {
+            $data = User::findOrFail($id);
+            $data->password = Hash::make('starfm@changeme');
+            $data->update();
+        }
+        abort(403, 'Unauthorized action.');
     }
 
     public function change_password()
@@ -235,19 +255,17 @@ class UsersController extends Controller
             'new_password' => 'required|confirmed|min:8|string'
         ]);
         $auth = Auth::user();
- 
+
         // The passwords matches
-        if (!Hash::check($request->get('old_password'), $auth->password)) 
-        {
+        if (!Hash::check($request->get('old_password'), $auth->password)) {
             return back()->with('error_password', "Current Password is Invalid");
         }
- 
+
         // Current password and new password same
-        if (strcmp($request->get('old_password'), $request->new_password) == 0) 
-        {
+        if (strcmp($request->get('old_password'), $request->new_password) == 0) {
             return redirect()->back()->with("error_password", "New Password cannot be same as your current password.");
         }
- 
+
         $user =  User::find($auth->id);
         $user->password =  Hash::make($request->new_password);
         $user->save();
@@ -261,9 +279,12 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
-        $data = User::findOrFail($id);
-        Storage::disk('public')->delete('users/profile/' . $data->profile_image);
-        Storage::disk('public')->delete('users/cover/' . $data->cover_image);
-        $data->delete();
+        if (auth()->user()->user_type == 1) {
+            $data = User::findOrFail($id);
+            Storage::disk('public')->delete('users/profile/' . $data->profile_image);
+            Storage::disk('public')->delete('users/cover/' . $data->cover_image);
+            $data->delete();
+        }
+        abort(403, 'Unauthorized action.');
     }
 }
